Fix: Move module to right submodule
